Add files via upload
A Python 2.0 project that manipulates sound. It can:

Change the audio file's speed (make it slower or faster).
Swap the two halves of an audio file and playing the resultant audio file.
Reverse the sound by redefining the new sample rate and playing it in reverse.
Change the volume of the sound by multiplying the number of samples.
Overlap the audio with a controllable static noise.
Take two audio files (each with different sample rates) and layers them into a new .wav file to play them both together. This function also has the capability of controlling the scale factor (volume) of each sound, so one sound will play a little louder than the other.
Add an echo to the sound file by adding a time delay.